Create Ofsted helper functions
